requirements.txt
discord.py==1.7.3
pytz==2022.1
